Viewservice
It passes all test cases. After the last test scenario, the viewservice
will be in “deadlock”. That’s because each view need to be acked by
primary to proceed to the next one. The last test case will kill the
primary and not let backup to take over. As a result, that view will
never be acked.