Primary/backup key/value service
Have some misunderstandings with the lab 2 requirement. It says client
should keep trying, but does not specify the time out between each try
or should the client talk to the viewservice and change its target
accordingly between each try. This commit will pass the last test, but
I am not confident that it is the wanted behavior.